rename: common폴더로 이동
