+++ src/components/common/DynamicBtn/DropDownBtn.tsx
@@ -1,38 +1,40 @@
-import type { BaseBtnProps, TextIconStatus } from "../../types/DynamicBtn";
-import useDynamicStatus from "../../hooks/useDynamicStatus";
+import type { BaseBtnProps, TextIconStatus } from "../../../types/DynamicBtn";
+import useDynamicStatus from "../../../hooks/useDynamicStatus";
 
-export default function TextIconBtnS({
+export default function DropDownBtn({
   children,
   disabled = false,
   onClick,
   type,
 }: BaseBtnProps) {
+  //상태 관리
+
   const { status, onMouseDown, onMouseLeave, onMouseUp } = useDynamicStatus(
     disabled,
     false,
   );
+
   const statusMap: Record<TextIconStatus, { bg?: string; icon: string }> = {
     clicked: {
-      icon: "/src/assets/icons/arrow_right.svg",
+      icon: "/src/assets/icons/arrow_down.svg",
     },
     pressing: {
-      bg: "bg-gy-100 ",
-      icon: "/src/assets/icons/arrow_right.svg",
+      bg: "bg-gy-100",
+      icon: "/src/assets/icons/arrow_down.svg",
     },
     default: {
-      icon: "/src/assets/icons/arrow_right.svg",
+      icon: "/src/assets/icons/arrow_down.svg",
     },
     disabled: {
       bg: "text-gy-400",
-      icon: "/src/assets/icons/arrow_rightG.svg",
+      icon: "/src/assets/icons/arrow_downG.svg",
     },
   };
-
   const { bg, icon } = statusMap[status as TextIconStatus];
 
   return (
     <button
-      className={`inline-flex pl-2 py-1 pr-1 gap-1 rounded-lg items-center  body-rg-400 ${bg} ${disabled ? "cursor-not-allowed" : "cursor-pointer"}`}
+      className={`inline-flex pr-[0.625rem] py-1 pl-4 gap-1 rounded-lg items-center header-h4  ${bg} ${disabled ? "cursor-not-allowed" : "cursor-pointer"}`}
       onMouseDown={onMouseDown}
       onMouseUp={onMouseUp}
       onMouseLeave={onMouseLeave}
@@ -40,7 +42,7 @@ export default function TextIconBtnS({
       type={type ? type : "button"}
     >
       {children}
-      <img src={icon} alt="버튼" className="size-4" />
+      <img src={icon} alt="지역 선택 메뉴열기" className=" size-4" />
     </button>
   );
 }
